Conflicts:
	src/smatrix.c

+++ src/smatrix.h
@@ -79,12 +79,12 @@ typedef struct {
 
 // --- public api
 smatrix_t* smatrix_open(const char* fname);
-uint64_t smatrix_get(smatrix_t* self, uint32_t x, uint32_t y);
-uint64_t smatrix_set(smatrix_t* self, uint32_t x, uint32_t y, uint64_t value);
-uint64_t smatrix_incr(smatrix_t* self, uint32_t x, uint32_t y, uint64_t value);
-uint64_t smatrix_decr(smatrix_t* self, uint32_t x, uint32_t y, uint64_t value);
-uint64_t smatrix_rowlen(smatrix_t* self, uint32_t x);
-uint64_t smatrix_getrow(smatrix_t* self, uint32_t x, uint64_t* ret, size_t ret_len);
+uint32_t smatrix_get(smatrix_t* self, uint32_t x, uint32_t y);
+uint32_t smatrix_set(smatrix_t* self, uint32_t x, uint32_t y, uint32_t value);
+uint32_t smatrix_incr(smatrix_t* self, uint32_t x, uint32_t y, uint32_t value);
+uint32_t smatrix_decr(smatrix_t* self, uint32_t x, uint32_t y, uint32_t value);
+uint32_t smatrix_rowlen(smatrix_t* self, uint32_t x);
+uint32_t smatrix_getrow(smatrix_t* self, uint32_t x, uint32_t* ret, size_t ret_len);
 void smatrix_close(smatrix_t* self);
 // ---
 
@@ -99,7 +99,7 @@ void smatrix_mfree(smatrix_t* self, uint64_t bytes);
 uint64_t smatrix_falloc(smatrix_t* self, uint64_t bytes);
 void smatrix_ffree(smatrix_t* self, uint64_t fpos, uint64_t bytes);
 
-void smatrix_rmap_init(smatrix_t* self, smatrix_rmap_t* rmap, uint64_t size);
+void smatrix_rmap_init(smatrix_t* self, smatrix_rmap_t* rmap, uint32_t size);
 smatrix_rmap_slot_t* smatrix_rmap_probe(smatrix_t* self, smatrix_rmap_t* rmap, uint32_t key);
 smatrix_rmap_slot_t* smatrix_rmap_insert(smatrix_t* self, smatrix_rmap_t* rmap, uint32_t key);
 void smatrix_rmap_resize(smatrix_t* self, smatrix_rmap_t* rmap);
